requirements.txt
# MCP Framework
mcp[cli]>=1.9.0

# LLM
anthropic>=0.42.0

# Database
aiosqlite>=0.20.0

# Document parsing
pdfplumber>=0.11.0
python-docx>=1.1.0

# Spreadsheet generation
openpyxl>=3.1.0

# Configuration
python-dotenv>=1.0.0

# Data validation
pydantic>=2.10.0

# HTTP transport
uvicorn>=0.34.0
httpx>=0.28.0
